+++ klum-ast-runtime/src/main/java/com/blackbuild/klum/ast/util/layer3/AutoCreationPhase.java
@@ -27,12 +27,19 @@ import com.blackbuild.klum.ast.process.DefaultKlumPhase;
 import com.blackbuild.klum.ast.process.VisitingPhaseAction;
 import com.blackbuild.klum.ast.util.*;
 import com.blackbuild.klum.ast.util.layer3.annotations.AutoCreate;
+import com.blackbuild.klum.ast.util.layer3.annotations.Cluster;
 import groovy.lang.Closure;
 
+import java.lang.annotation.Annotation;
+import java.lang.reflect.AnnotatedElement;
 import java.lang.reflect.Field;
+import java.lang.reflect.Method;
+import java.lang.reflect.Type;
 import java.util.Map;
+import java.util.Optional;
+import java.util.function.Predicate;
 
-import static com.blackbuild.klum.ast.util.DslHelper.isInstantiable;
+import static com.blackbuild.klum.ast.util.DslHelper.*;
 import static java.lang.String.format;
 
 public class AutoCreationPhase extends VisitingPhaseAction {
@@ -43,19 +50,20 @@ public class AutoCreationPhase extends VisitingPhaseAction {
 
     @Override
     public void visit(String path, Object element, Object container, String nameOfFieldInContainer) {
-        ClusterModel.getFieldsAnnotatedWith(element, AutoCreate.class)
-                .entrySet()
-                .stream()
+        ClusterModel.getPropertiesStream(element, Object.class)
                 .filter(entry -> entry.getValue() == null)
-                .forEach(entry -> autoCreate(element, entry.getKey()));
+                .map(pv -> ClusterModel.getField(element, pv.getName()))
+                .filter(Optional::isPresent)
+                .map(Optional::get)
+                .filter(field -> field.isAnnotationPresent(AutoCreate.class))
+                .forEach(field -> autoCreate(element, field, field.getAnnotation(AutoCreate.class)));
+
+        autoCreateClusterFields(element);
 
         LifecycleHelper.executeLifecycleMethods(KlumInstanceProxy.getProxyFor(element), AutoCreate.class);
     }
 
-    private void autoCreate(Object element, String fieldName) {
-        @SuppressWarnings("OptionalGetWithoutIsPresent")
-        Field field = DslHelper.getField(element.getClass(), fieldName).get();
-        AutoCreate autoCreate = field.getAnnotation(AutoCreate.class);
+    private void autoCreate(Object element, Field field, AutoCreate autoCreate) {
         Map<String, Object> values = ClosureHelper.invokeClosure(autoCreate.value());
 
         String key = autoCreate.key();
@@ -63,24 +71,54 @@ public class AutoCreationPhase extends VisitingPhaseAction {
             key = null;
         // TODO: Validation AST
         if (key == null && DslHelper.isKeyed(field.getType()))
-            throw new KlumSchemaException(format("AutoCreate annotation for field '%s' is missing a 'key' field.", fieldName));
+            throw new KlumSchemaException(format("AutoCreate annotation for field '%s' is missing a 'key' field.", field.getName()));
         else if (key != null && !DslHelper.isKeyed(field.getType()))
-            throw new KlumSchemaException(format("AutoCreate annotation for field '%s' has a key field, but annotated type '%s' is not keyed", fieldName, field.getType().getName()));
+            throw new KlumSchemaException(format("AutoCreate annotation for field '%s' has a key field, but annotated type '%s' is not keyed", field.getName(), field.getType().getName()));
 
         Class<?> type = autoCreate.type();
         if (type.equals(Object.class)) {
             if (field.getType().equals(Closure.class)) return;
             if (!isInstantiable(field.getType()))
-                throw new KlumSchemaException(format("AutoCreate annotation for abstract typed field '%s' is missing a 'type' field.", fieldName));
+                throw new KlumSchemaException(format("AutoCreate annotation for abstract typed field '%s' is missing a 'type' field.", field.getName()));
             type = field.getType();
         } else if (!field.getType().isAssignableFrom(type)) {
             throw new KlumSchemaException(
-                    format("AutoCreate annotation for field '%s' sets type '%s' which is no subtype of the field's type (%s)", fieldName, type, field.getType()));
+                    format("AutoCreate annotation for field '%s' sets type '%s' which is no subtype of the field's type (%s)", field.getName(), type, field.getType()));
         }
 
         Object autoCreated = FactoryHelper.create(type, values, key, null);
 
-        KlumInstanceProxy.getProxyFor(element).setSingleField(fieldName, autoCreated);
+        KlumInstanceProxy.getProxyFor(element).setSingleField(field.getName(), autoCreated);
     }
+
+    private void autoCreateClusterFields(Object element) {
+
+        ClusterModel.getMethodsAnnotatedWithStream(element, Map.class, Cluster.class)
+                .filter(clusterField -> clusterField.isAnnotationPresent(AutoCreate.class))
+                .forEach(clusterMethod -> autoCreateElementsForCluster(element, clusterMethod));
+    }
+
+    private void autoCreateElementsForCluster(Object element, Method clusterMethod) {
+        Cluster cluster = clusterMethod.getAnnotation(Cluster.class);
+        Class<? extends Annotation> filterAnnotation = cluster.value();
+        Predicate<AnnotatedElement> clusterFilter = filterAnnotation != Cluster.Undefined.class ? elementToCheck -> elementToCheck.isAnnotationPresent(filterAnnotation) : elementToCheck -> true;
+
+        Type elementType = getElementType(clusterMethod.getGenericReturnType());
+        if (!(elementType instanceof Class))
+            throw new KlumSchemaException(format("Cluster annotation on method '%s', whose element generics is no class (%s)", clusterMethod.getName(), elementType.getTypeName()));
+
+        if (!isDslType(elementType))
+            return;
+
+        AutoCreate autoCreate = clusterMethod.getAnnotation(AutoCreate.class);
+
+        ClusterModel.getPropertiesStream(element, (Class<?>) elementType, clusterFilter)
+                .filter(propertyValue -> propertyValue.getValue() == null)
+                .map(pv -> ClusterModel.getField(element, pv.getName()))
+                .filter(Optional::isPresent)
+                .map(Optional::get)
+                .forEach(property -> autoCreate(element, property, autoCreate));
+    }
+
 
 }