+++ klum-ast-runtime/src/main/java/com/blackbuild/klum/ast/util/layer3/ClusterModel.java
@@ -295,6 +295,10 @@ public class ClusterModel {
                 .collect(toList());
     }
 
+    public static Optional<Field> getField(Object container, String fieldName) {
+        return getField(container.getClass(), fieldName);
+    }
+
     public static Optional<Field> getField(Class<?> containerType, String fieldName) {
         while (containerType != null) {
             Optional<Field> field = Arrays.stream(containerType.getDeclaredFields()).filter(it -> it.getName().equals(fieldName)).findFirst();
@@ -409,11 +413,12 @@ public class ClusterModel {
      */
     public static Map<String, Object> getMethodBasedValues(Object object, Class<? extends Annotation> annotation, Class<?> returnType) {
         return getMethodsAnnotatedWithStream(object, returnType, annotation)
+                .filter(Predicate.not(ClusterModel::isGetter))
                 .collect(toMap(Method::getName, it -> InvokerHelper.invokeMethod(object, it.getName(), null)));
     }
 
     /**
-     * Returns all parameterless methods of the given Return type (possibly filtered).
+     * Returns all parameterless methods (possibly filtered).
      * @param object The object whose methods should be returned
      * @return A list of matching methods
      */
@@ -423,13 +428,14 @@ public class ClusterModel {
     }
 
     @NotNull
-    private static Stream<Method> getMethodsAnnotatedWithStream(Object object, Class<?> returnType, Class<? extends Annotation> annotation) {
+    static Stream<Method> getMethodsAnnotatedWithStream(Object object, Class<?> returnType, Class<? extends Annotation> annotation) {
         return Arrays.stream(object.getClass().getMethods())
-                .filter(it -> it.getParameterCount() == 0)
                 .filter(it -> returnType.isAssignableFrom(it.getReturnType()))
-                .filter(it -> !it.getName().startsWith("get"))
-                .filter(it -> !it.getName().startsWith("is"))
                 .filter(it -> it.isAnnotationPresent(annotation));
+    }
+
+    private static boolean isGetter(Method method) {
+        return method.getParameterCount() == 0 && (method.getName().startsWith("get") || method.getName().startsWith("is"));
     }
 
 }