+++ CHANGES.md
@@ -48,6 +48,7 @@
 - @LinkTo now correctly handles empty collections/maps as target
 - Allow a custom key-provider function for `createFrom(URL)` and `createFrom(File)` 
 - `@Cluster` can also be placed on fields, which will be converted into a getter method (see [#366](https://github.com/klum-dsl/klum-ast/issues/366))
+- `@Cluster` can be combined with `@AutoCreate` to auto create all members of the annotated cluster (see [#363](https://github.com/klum-dsl/klum-ast/issues/363))
 
 ## Deprecations (see [Migration](https://github.com/klum-dsl/klum-ast/wiki/Migration)):
   - The `@Validation` annotation is deprecated. Use `@Validate` on class level instead.